Oops: get_num_samples return int

+++ src/spikeinterface/preprocessing/decimate.py
@@ -114,7 +114,7 @@ class DecimateRecordingSegment(BaseRecordingSegment):
     def get_num_samples(self):
         parent_n_samp = self._parent_segment.get_num_samples()
         assert self._decimation_offset < parent_n_samp  # Sanity check (already enforced). Formula changes otherwise
-        return np.ceil((parent_n_samp - self._decimation_offset) / self._decimation_factor)
+        return int(np.ceil((parent_n_samp - self._decimation_offset) / self._decimation_factor))
 
     def get_traces(self, start_frame, end_frame, channel_indices):
         if start_frame is None:
